+++ main.lua
@@ -1,7 +1,10 @@
 local game = require("game")
 local fsm = require("fsm")
 local home = require("home")
+
+-- Uncomment to test the state immediately.
 -- local ep1 = require("ep1")
+-- local ep2 = require("ep2")
 
 function love.load()
     love.graphics.setDefaultFilter("nearest", "nearest")  -- Nearest neighbor filtering for pixel art.
@@ -9,7 +12,10 @@ function love.load()
 
     game.init()
     fsm.push(home)
+
+    -- Uncomment to test the state immediately.
     -- fsm.push(ep1)
+    -- fsm.push(ep2)
 end
 
 function love.update(dt)